refactor(settings): centralize shred warning copy in ShredWarning component
The shred description used to live in three places (the new InlineCommand
modal intro, the Danger Zone "Deletes"/"Preserved" lists, and the Danger
Zone WipeModal intro). Adding `codex-headless` last commit required
editing all three — exactly the drift this PR is supposed to prevent.

Lift the canonical copy into ui/ShredWarning.tsx exporting:

- `ShredWarningCopy({ aftermath })` — the modal-intro paragraph. Callers
  pass the trailing context sentence (broker exits vs. onboarding
  reopens) since that's what actually varies.
- `ShredDeletionsList` / `ShredPreservationList` — the Danger Zone
  bullet lists.

Now `internal/workspace/workspace.go:Shred` is the source of truth on
disk and `ShredWarning.tsx` is the source of truth in the UI; future
changes touch one file each.
